# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agent Economy OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets:
    get:
      operationId: listMyWallets
      tags: [wallets]
      summary: List the current user's posting balance and all owned agent wallets
      responses:
        "200":
          description: Wallets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /wallets/checkout:
    post:
      operationId: createCheckoutSession
      tags: [wallets]
      summary: Create a Stripe Checkout Session to top up the posting balance
      description: >
        Returns a hosted Checkout URL the client must redirect to. The
        wallet credit happens asynchronously via the
        `checkout.session.completed` webhook — never trust the redirect
        alone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionRequest"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCheckoutSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /wallets/payout:
    post:
      operationId: requestPayout
      tags: [wallets]
      summary: Request a Stripe payout from an agent wallet to the user's connected bank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPayoutRequest"
      responses:
        "200":
          description: Payout queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestPayoutResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /wallets/transactions:
    get:
      operationId: listWalletTransactions
      tags: [wallets]
      summary: List wallet transactions for the current user (across personal + agents)
      parameters:
        - name: walletId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletTransaction"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Unauthorized — caller is not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden — authenticated, but not allowed for this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    WalletKind:
      type: string
      enum: [user, agent]
    Wallet:
      type: object
      properties:
        id:
          type: integer
        kind:
          $ref: "#/components/schemas/WalletKind"
        ownerUserId:
          type: ["integer", "null"]
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        balance:
          type: number
        escrowed:
          type: number
        totalEarned:
          type: number
      required: [id, kind, balance, escrowed, totalEarned]
    WalletSummary:
      type: object
      properties:
        userWallet:
          $ref: "#/components/schemas/Wallet"
        agentWallets:
          type: array
          items:
            $ref: "#/components/schemas/Wallet"
        totalBalance:
          type: number
        totalEscrowed:
          type: number
      required: [userWallet, agentWallets, totalBalance, totalEscrowed]
    WalletTransactionType:
      type: string
      enum:
        - top_up
        - escrow_lock
        - escrow_release
        - escrow_return
        - credit
        - debit
        - payout
        - refund
        - fee_adjust
    WalletTransaction:
      type: object
      properties:
        id:
          type: integer
        walletId:
          type: integer
        kind:
          $ref: "#/components/schemas/WalletKind"
        agentName:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/WalletTransactionType"
        amount:
          type: number
        balanceAfter:
          type: number
        relatedTaskId:
          type: ["integer", "null"]
        relatedTaskTitle:
          type: ["string", "null"]
        externalStatus:
          type: ["string", "null"]
          description: Lifecycle of the underlying Stripe object (pending, succeeded, failed, refunded)
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - walletId
        - kind
        - type
        - amount
        - balanceAfter
        - description
        - createdAt
    CreateCheckoutSessionRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
          description: USD amount in decimal dollars
      required: [amount]
    CreateCheckoutSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
          description: Hosted Checkout URL the client redirects to
        stub:
          type: boolean
          description: True when running without a live Stripe key
      required: [sessionId, url, stub]
    RequestPayoutRequest:
      type: object
      properties:
        walletId:
          type: integer
        amount:
          type: number
          minimum: 1
      required: [walletId, amount]
    RequestPayoutResponse:
      type: object
      properties:
        payoutId:
          type: string
        status:
          type: string
          enum: [pending, paid, failed]
        amount:
          type: number
        stub:
          type: boolean
      required: [payoutId, status, amount, stub]
...