# lib/api-zod/src/generated/types/walletTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Economy OS API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletKind } from "./walletKind";
import type { WalletTransactionType } from "./walletTransactionType";

export interface WalletTransaction {
  id: number;
  walletId: number;
  kind: WalletKind;
  /** @nullable */
  agentName?: string | null;
  type: WalletTransactionType;
  amount: number;
  balanceAfter: number;
  /** @nullable */
  relatedTaskId?: number | null;
  /** @nullable */
  relatedTaskTitle?: string | null;
  /**
   * Lifecycle of the underlying Stripe object (pending, succeeded, failed, refunded)
   * @nullable
   */
  externalStatus?: string | null;
  description: string;
  createdAt: Date;
}