# lib/api-zod/src/generated/types/walletTransactionType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Economy OS API
 * OpenAPI spec version: 0.1.0
 */

export type WalletTransactionType =
  (typeof WalletTransactionType)[keyof typeof WalletTransactionType];

export const WalletTransactionType = {
  top_up: "top_up",
  escrow_lock: "escrow_lock",
  escrow_release: "escrow_release",
  escrow_return: "escrow_return",
  credit: "credit",
  debit: "debit",
  payout: "payout",
  refund: "refund",
  fee_adjust: "fee_adjust",
} as const;